requirements.txt
openpyxl==3.1.5
pandas==2.3.3
azure-search-documents==11.6.0
azure-identity==1.19.0
python-dotenv==1.1.1
openai==2.5.0
streamlit==1.50.0
